+++ src/main/java/com/shaft/gui/image/GifSequenceWriter.java
@@ -1,4 +1,4 @@
-package com.shaft.image;
+package com.shaft.gui.image;
 
 //
