+++ src/main/java/com/shaft/validation/Verifications.java
@@ -6,12 +6,12 @@ import org.testng.Assert;
 import org.testng.ITestResult;
 import org.testng.Reporter;
 
-import com.shaft.browser.BrowserActions;
-import com.shaft.element.ElementActions;
-import com.shaft.image.ScreenshotManager;
-import com.shaft.io.FileActions;
-import com.shaft.io.ReportManager;
-import com.shaft.support.JavaActions;
+import com.shaft.cli.FileActions;
+import com.shaft.gui.browser.BrowserActions;
+import com.shaft.gui.element.ElementActions;
+import com.shaft.gui.image.ScreenshotManager;
+import com.shaft.tools.io.ReportManager;
+import com.shaft.tools.support.JavaActions;
 
 public class Verifications {
