+++ src/test/java/poms/GoogleSearch.java
@@ -3,8 +3,8 @@ package poms;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 
-import com.shaft.browser.BrowserActions;
-import com.shaft.element.ElementActions;
+import com.shaft.gui.browser.BrowserActions;
+import com.shaft.gui.element.ElementActions;
 import com.shaft.validation.Assertions;
 import com.shaft.validation.Verifications;
