+++ src/test/java/poms/GoogleSearchResults.java
@@ -3,7 +3,7 @@ package poms;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 
-import com.shaft.element.ElementActions;
+import com.shaft.gui.element.ElementActions;
 import com.shaft.validation.Assertions;
 import com.shaft.validation.Verifications;
