+++ src/test/java/testPackage01/Test_FileActions.java
@@ -1,0 +1,13 @@
+package testPackage01;
+
+import org.testng.annotations.Test;
+
+import com.shaft.cli.FileActions;
+import com.shaft.tools.io.ReportManager;
+
+public class Test_FileActions {
+    @Test
+    public void f() {
+	ReportManager.log(FileActions.listFilesInDirectory("/home/"));
+    }
+}
