+++ src/test/java/testPackage01/Test_checksum.java
@@ -2,8 +2,8 @@ package testPackage01;
 
 import org.testng.annotations.Test;
 
+import com.shaft.cli.FileActions;
 import com.shaft.cli.TerminalActions;
-import com.shaft.io.FileActions;
 import com.shaft.validation.Assertions;
 
 public class Test_checksum {
