+++ src/test/java/testPackage01/Test_fullPageScreenshotWithHeader.java
@@ -5,8 +5,8 @@ import org.openqa.selenium.WebDriver;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
-import com.shaft.browser.BrowserActions;
-import com.shaft.browser.BrowserFactory;
+import com.shaft.gui.browser.BrowserActions;
+import com.shaft.gui.browser.BrowserFactory;
 import com.shaft.validation.Assertions;
 
 public class Test_fullPageScreenshotWithHeader {
