+++ src/test/java/testPackage01/Test_google.java
@@ -4,8 +4,8 @@ import org.openqa.selenium.WebDriver;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
-import com.shaft.browser.BrowserFactory;
-import com.shaft.io.ExcelFileManager;
+import com.shaft.gui.browser.BrowserFactory;
+import com.shaft.tools.io.ExcelFileManager;
 
 import poms.GoogleSearch;
 import poms.GoogleSearchResults;
