+++ src/test/java/testPackage01/Test_iFrame.java
@@ -5,9 +5,9 @@ import org.openqa.selenium.WebDriver;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
-import com.shaft.browser.BrowserActions;
-import com.shaft.browser.BrowserFactory;
-import com.shaft.element.ElementActions;
+import com.shaft.gui.browser.BrowserActions;
+import com.shaft.gui.browser.BrowserFactory;
+import com.shaft.gui.element.ElementActions;
 import com.shaft.validation.Verifications;
 
 public class Test_iFrame {
