+++ src/test/java/unitTests/tests_validations_assertions.java
@@ -5,10 +5,10 @@ import org.openqa.selenium.WebDriver;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
-import com.shaft.browser.BrowserActions;
-import com.shaft.browser.BrowserFactory;
-import com.shaft.element.ElementActions;
-import com.shaft.io.ReportManager;
+import com.shaft.gui.browser.BrowserActions;
+import com.shaft.gui.browser.BrowserFactory;
+import com.shaft.gui.element.ElementActions;
+import com.shaft.tools.io.ReportManager;
 import com.shaft.validation.Assertions;
 
 public class tests_validations_assertions {
